fix(seller incomes): fixing seller incomes data `undefined` bug
